progress commit: tracing mal log-in message
It is now clear, that the myAnimeList log-in is called in a deferred
manner in the myanimelist module. This leads to some confusing
conversion errors, as there are suddenly two different deferring models
working in the background (gtk+ and twisted callbacks).
The next step is to make the twisted identify function non-deferred and
integrate it in the gtk/main module.