Rewrite of the myanimelist module phase 1/2
The login and fetching of anime lists is now completely rewritten,
maintainable and nice. The thing that does not work now is the
committing of local data. Would be nice to have the entry update
data from mal, but it is not in the XML. Wrote a question to Xinil for
this and await answer.

The local data representation is completely changed, so the main
application does not work now.

Twisted was removed form the myanimelist module. The reason is, that it
is a total overkill for this application. We only need http fetching and
posting, no need for twisted (which is much more suited if we wanted
to write a webserver). Also the twisted documentation is horrible.

Next up is the rewrite of the user interface.