myanimelist dirty implementation
As Xinil added the my_last_updated field to the XML, the data sync
module was extended by sync capabilities.

Note: the new functions are not tested and must be debugged as a next step.
sync_mal will be morphed to a class and will get a shelf based data
storage routine for datafile creation/loading on init and writing on
local update and sync.