myanimelist.py interface update
Replaced the sync_mal function with the anime_data class, which also
handles the local data file. It has a fetch and sync method and a
database attribute. This makes it easy to sync the local file, app data
and remote data at the same time. This approach is also much more readable.