Added document gen. system and changed interface integration name.
The generate_html_docs.py now auto-generates html doc files in the doc
directory of the code via pydoc. Please don't add the actual html files
to the repository, as they would cause a bunch of merging problems.

Also renamed main.py to gtkmain.py as we are thinking about using
exchangeable gui modules and this makes it clear which main we are
talking about.