GUI progress commit.
This commit features the first working callback mechanics for the
TreeView's that display the anime data (the panels in the middle) and
which are used for manual data manipulation (episode/status/score change).

...

And for the record, here is the club comment of the day on the project
status.

Development is progressing well. The GUI is a bit more tedious than
I assumed, but I'm through with the hardest part of the mechanics:
the pygtk TreeView (the panels that display the anime lists and
through witch they can be manipulated). Whoever programmed that
part of the library was obviously a sadistic daemon.

There is still quite some callback and data logic, auxiliary
functionality (like the preferences window) and
playtracker/recognition subsystem integration and some clean-up
to be done before release.

Necrotex is also having good progress on the recognition engine.
Once I'm done with the GUI to on a satisfactory state I'll test
and integrate that one, which will be the last item of our pre-beta
todo list. Stay tuned..