GUI Progress -- cleanup and documentation
To make things nice and tidy before getting into further implementation
detail I made a small structural cleanup (imports are now done nice
the python handbook way, treeview data interface setup, subsystem wrapper
global).

I also reviewed all the documentation and extended it where it needed it.
I was also a bit in a funny mood, but the information is still sensible,
correct and nearly in the right volume to easily scan through the code and
understand what is going on.