Closes gh-13 . Fixed missing event handlers and added about logo.
Seems like during the name conversion of the ui xml the event handlers
got deleted. This happend due to a different program version of glade 3.
Don't ask me, it just ate the event handlers, and now I have put them back.

And as I was playing around with the ui, I also put in the new big glass
looking logo.