Added customized wx spin control class.
Now this is the first functional widget for the upcoming wx interface.
It is a combination of TextCtrl and SpinCtrl with nifty mouse over
effect for good looks and efficient data manipulation. More comments will
be added later.

Run it with python to see a small test with 5 spin custom spin ctrls.