New configuration architecture
Allow command line arguments to override the configuration file,
which is very nice for testing and debugging.

Start to approach an MVC architecture.

(ac_config can be seen as the configuration controller for two
model objects: DefaultConfig and the return value of options())

Get rid of the ugly rundat pseudo-global in favour of explicitly
passing needed data to the GUI controller.