Added title index
Not used by anything yet. Could be used for all kinds of sites.

Easier to understand than N-gram, but fails when not a single word matches,
like for wildly misspelled one-word titles.

Faster than the recognizing module (pure difflib). Takes 0.1 seconds to find
something on the crono22 list, as opposed to 6 seconds.

Also searches synonyms, which recognizing does not.